==> 3 - TreinamentoAlgoritmos/2. Tratamento do Texto.ipynb <==
# Tratamento do Texto

import pickle
import pandas as pd
import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import chi2
import numpy as np
import re

with open('StackOverflow_dataset.pickle', 'rb') as data:
    df = pickle.load(data)

df.head()

#df.loc[1]['body']
df['body']=df['title']+df['body']

df['body']

## Limpar o Texto

df['Content_Parsed_1'] = df['body'].str.replace("\r", " ")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("\n", " ")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("    ", " ")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("*", "")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("<p>", "")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("</p>", "")
df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace("/", "")
# df['Content_Parsed_1'] = df['Content_Parsed_1'].replace(regex=r'\<.*>', value='')



df['Content_Parsed_1'] = df['Content_Parsed_1'].str.replace('"', '')

### Maiúsculo para minúsculo

df['Content_Parsed_2'] = df['Content_Parsed_1'].str.lower()

### Retirando pontos

punctuation_signs = list("?:!.,;")
df['Content_Parsed_3'] = df['Content_Parsed_2']

for punct_sign in punctuation_signs:
    df['Content_Parsed_3'] = df['Content_Parsed_3'].str.replace(punct_sign, '')

### Removendo pronome de posse

df['Content_Parsed_4'] = df['Content_Parsed_3'].str.replace("'s", "")

### Stemming e Lemmatization

Desde que aretirada de caracteres feita anteriormente, pode produzir palavras que não existem, usei o processo de lemmatization. 

nltk.download('punkt')
print("------------------------------------------------------------")
nltk.download('wordnet')

wordnet_lemmatizer = WordNetLemmatizer()

nrows = len(df)
lemmatized_text_list = []

for row in range(0, nrows):
    
    lemmatized_list = []
    
    text = df.loc[row]['Content_Parsed_4']
    text_words = str(text).split(" ")

    for word in text_words:
        lemmatized_list.append(wordnet_lemmatizer.lemmatize(word, pos="v"))
        
    lemmatized_text = " ".join(lemmatized_list)
    
    lemmatized_text_list.append(lemmatized_text)

df['Content_Parsed_5'] = lemmatized_text_list

Pronomes

nltk.download('stopwords')

stop_words = list(stopwords.words('english'))

df['Content_Parsed_6'] = df['Content_Parsed_5']

for stop_word in stop_words:

    regex_stopword = r"\b" + stop_word + r"\b"
    df['Content_Parsed_6'] = df['Content_Parsed_6'].str.replace(regex_stopword, '')

list_columns = ["labels","body", "Content_Parsed_6"] #pegar o titulo 
df = df[list_columns]

df = df.rename(columns={'Content_Parsed_6': 'Content_Parsed'})

df.head()

## Mapeando as categorias

category_codes = {
    'bug': 0,
    'enhancement': 1,
    'question': 2,
}
#categorias vou por aqui

df['Category_Code'] = df['labels']
df = df.replace({'Category_Code':category_codes})

## Divisão de treino e teste

X_train, X_test, y_train, y_test = train_test_split(df['Content_Parsed'], 
                                                    df['Category_Code'], 
                                                    test_size=0.30, 
                                                    random_state=10)

y_train

## Representação do texto

ngram_range = (1,2)
min_df = 10
max_df = 1.
max_features = 300

tfidf = TfidfVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,
                        norm='l2',
                        sublinear_tf=True)
                        
features_train = tfidf.fit_transform(X_train).toarray()
labels_train = y_train
print(features_train.shape)

features_test = tfidf.transform(X_test).toarray()
labels_test = y_test
print(features_test.shape)

from sklearn.feature_selection import chi2
import numpy as np

for Product, category_id in sorted(category_codes.items()):
    features_chi2 = chi2(features_train, labels_train == category_id)
    indices = np.argsort(features_chi2[0])
    feature_names = np.array(tfidf.get_feature_names())[indices]
    unigrams = [v for v in feature_names if len(v.split(' ')) == 1]
    bigrams = [v for v in feature_names if len(v.split(' ')) == 2]
    print("# '{}' labels:".format(Product))
    print("  . Most correlated unigrams:\n. {}".format('\n. '.join(unigrams[-5:])))
    print("  . Most correlated bigrams:\n. {}".format('\n. '.join(bigrams[-2:])))
    print("")


# X_train
with open('Pickles/X_train.pickle', 'wb') as output:
    pickle.dump(X_train, output)
    
# X_test    
with open('Pickles/X_test.pickle', 'wb') as output:
    pickle.dump(X_test, output)
    
# y_train
with open('Pickles/y_train.pickle', 'wb') as output:
    pickle.dump(y_train, output)
    
# y_test
with open('Pickles/y_test.pickle', 'wb') as output:
    pickle.dump(y_test, output)
    
# df
with open('Pickles/df.pickle', 'wb') as output:
    pickle.dump(df, output)
    
# features_train
with open('Pickles/features_train.pickle', 'wb') as output:
    pickle.dump(features_train, output)

# labels_train
with open('Pickles/labels_train.pickle', 'wb') as output:
    pickle.dump(labels_train, output)

# features_test
with open('Pickles/features_test.pickle', 'wb') as output:
    pickle.dump(features_test, output)

# labels_test
with open('Pickles/labels_test.pickle', 'wb') as output:
    pickle.dump(labels_test, output)
    
# TF-IDF object
with open('Pickles/tfidf.pickle', 'wb') as output:
    pickle.dump(tfidf, output)
