==> 3 - TreinamentoAlgoritmos/8. BERT.ipynb <==
pip install ktrain

import ktrain
from ktrain import text
import pandas as pd

from google.colab import drive
drive.mount('/content/drive')

import os
os.chdir("/content/drive/MyDrive/algoritmos/")

dados = pd.read_csv("9.FestText/all.csv")

dados

(x_train, y_train) , (x_test, y_test), preproc = text.texts_from_df(dados, text_column="new_text", preprocess_mode="bert", label_columns=["new_labels"], val_pct=0.3)

classifier = text.text_classifier(name="bert", train_data=(x_train, y_train), preproc=preproc)

learner = ktrain.get_learner(classifier, train_data=(x_train, y_train), val_data=(x_test, y_test), batch_size=8)

learner.lr_find()

learner.fit_onecycle(2e-5, 10)

learner.evaluate((x_test, y_test))

predictor = ktrain.get_predictor(learner.model, preproc)

predictor.get_classes()
