==> 3 - TreinamentoAlgoritmos/9. Best Model Selection.ipynb <==
# Best Model Selection

import pickle
import pandas as pd

path_pickles = "Models/"

list_pickles = [
    "df_models_gbc.pickle",
    "df_models_lrc.pickle",
    "df_models_mnbc.pickle",
    "df_models_rfc.pickle",
    "df_models_svc.pickle"
#se adicionar ou trocar, mudar essa parte do código
]

df_summary = pd.DataFrame()

for pickle_ in list_pickles:
    
    path = path_pickles + pickle_
    
    with open(path, 'rb') as data:
        df = pickle.load(data)

    df_summary = df_summary.append(df)

df_summary = df_summary.reset_index().drop('index', axis=1)

df_summary

df_summary.sort_values('Test Set Accuracy', ascending=False)
